+++ src/master/contentlayout.js
@@ -11,7 +11,7 @@ const ContentLayout = ({ children }) => {
   return (
     <>
         <div className={`${category !== 'home' ? `bg-${theme} w-100 h-auto px-8 md:px-8 lg:px-16` : 'hidden' } `}>
-            <div className='bg-slate-50 pt-24 px-8'>
+            <div className='bg-slate-50 pt-24'>
             { children } 
             </div>
         </div>